implement type changes
