+++ packages/common-all/src/types/index.ts
@@ -14,6 +14,7 @@ export * from "./intermediateConfigs";
 export * from "./compat";
 export * from "./editor";
 export * from "./lookup";
+export * from "./unified";
 
 export type Stage = "dev" | "prod" | "test";
 export type DEngineQuery = {
