+++ packages/common-all/src/types/typesv2.ts
@@ -15,35 +15,7 @@ import { IntermediateDendronConfig } from "./intermediateConfigs";
 import { VSRange } from "./compat";
 import { Decoration, Diagnostic } from ".";
 import type { NoteFNamesDict } from "../utils";
-
-export enum DendronASTDest {
-  MD_ENHANCED_PREVIEW = "MD_ENHANCED_PREVIEW",
-  MD_REGULAR = "MD_REGULAR",
-  MD_DENDRON = "MD_DENDRON",
-  HTML = "HTML",
-}
-
-/**
- * If processor should run in an alternative flavor
- */
-export enum ProcFlavor {
-  /**
-   * No special processing
-   */
-  REGULAR = "REGULAR",
-  /**
-   * Apply publishing rules
-   */
-  PUBLISHING = "PUBLISHING",
-  /**
-   * Apply preview rules
-   */
-  PREVIEW = "PREVIEW",
-  /**
-   * Apply hover preview rules (used for the preview when hovering over a link)
-   */
-  HOVER_PREVIEW = "HOVER_PREVIEW",
-}
+import { DendronASTDest, ProcFlavor } from "./unified";
 
 export enum ResponseCode {
   OK = 200,
